+++ projects/angular-showcase/src/app/business/business/business.component.ts
@@ -3,9 +3,18 @@ import { Component } from '@angular/core';
 
 import { ExampleProvider } from '../../shared/example-provider';
 import { DialogShowcaseComponent } from '../business-examples/dialog-showcase/dialog-showcase.component';
+import { ClosableNotificationComponent } from '../business-examples/notification-showcase/closable-notification/closable-notification.component';
+import { CustomIconNotificationComponent } from '../business-examples/notification-showcase/custom-icon-notification/custom-icon-notification.component';
+import { JumpmarkNotificationComponent } from '../business-examples/notification-showcase/jumpmark-notification/jumpmark-notification.component';
+import { SimpleNotificationComponent } from '../business-examples/notification-showcase/simple-notification/simple-notification.component';
+import { ToastNotificationComponent } from '../business-examples/notification-showcase/toast-notification/toast-notification.component';
 import { PaginationShowcaseComponent } from '../business-examples/pagination-showcase/pagination-showcase.component';
 import { SimpleContextmenuComponent } from '../business-examples/simple-contextmenu/simple-contextmenu.component';
 import { SkippableProcessflowComponent } from '../business-examples/skippable-processflow/skippable-processflow.component';
+import { GroupedColumnsTableComponent } from '../business-examples/table-showcase/grouped-columns-table/grouped-columns-table.component';
+import { GroupedRowsTableComponent } from '../business-examples/table-showcase/grouped-rows-table/grouped-rows-table.component';
+import { SimpleTableComponent } from '../business-examples/table-showcase/simple-table/simple-table.component';
+import { SortableTableComponent } from '../business-examples/table-showcase/sortable-table/sortable-table.component';
 import { TabsShowcaseComponent } from '../business-examples/tabs-showcase/tabs-showcase.component';
 import { TooltipShowcaseComponent } from '../business-examples/tooltip-showcase/tooltip-showcase.component';
 import { UsermenuShowcaseComponent } from '../business-examples/usermenu-showcase/usermenu-showcase.component';
@@ -33,9 +42,11 @@ export class BusinessComponent implements ExampleProvider {
   layoutComponents = {
     accordion: 'Accordion',
     pagination: 'Pagination',
+    notification: 'Notification',
     processflow: 'Processflow',
     tabs: 'Tabs',
-    usermenu: 'Usermenu'
+    usermenu: 'Usermenu',
+    table: 'Table'
   };
   buttonAndIndicatorComponents = {
     button: 'Button',
@@ -66,6 +77,19 @@ export class BusinessComponent implements ExampleProvider {
     },
     pagination: {
       'pagination-showcase': new ComponentPortal(PaginationShowcaseComponent)
+    },
+    notification: {
+      'simple-notification': new ComponentPortal(SimpleNotificationComponent),
+      'custom-icon-notification': new ComponentPortal(CustomIconNotificationComponent),
+      'toast-notification': new ComponentPortal(ToastNotificationComponent),
+      'jumpmark-notification': new ComponentPortal(JumpmarkNotificationComponent),
+      'closable-notification': new ComponentPortal(ClosableNotificationComponent)
+    },
+    table: {
+      'simple-table': new ComponentPortal(SimpleTableComponent),
+      'grouped-columns-table': new ComponentPortal(GroupedColumnsTableComponent),
+      'grouped-rows-table': new ComponentPortal(GroupedRowsTableComponent),
+      'sortable-table': new ComponentPortal(SortableTableComponent)
     }
   };