+++ projects/angular-showcase/src/app/icons/icons-routing.module.ts
@@ -3,13 +3,11 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { MarkdownViewerComponent } from '../shared/markdown-viewer/markdown-viewer.component';
 
-import { IconSearchComponent } from './icon-search/icon-search.component';
-import { IconViewerComponent } from './icon-viewer/icon-viewer.component';
 import { IconsComponent } from './icons/icons.component';
 
 const routes: Routes = [
   {
-    path: 'icons',
+    path: '',
     component: IconsComponent,
     data: { library: 'angular-icons' },
     children: [
@@ -25,11 +23,19 @@ const routes: Routes = [
       },
       {
         path: 'search',
-        component: IconSearchComponent
+        redirectTo: 'components'
       },
       {
-        path: 'components/:id',
-        component: IconViewerComponent
+        path: 'components',
+        loadChildren: () =>
+          import('./icon-components/icon-components.module').then(m => m.IconComponentsModule)
+      },
+      {
+        path: 'experimental',
+        loadChildren: () =>
+          import('./experimental-icons/experimental-icons.module').then(
+            m => m.ExperimentalIconsModule
+          )
       }
     ]
   }