+++ projects/angular-showcase/src/app/public/public-examples/select-forms/select-forms.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'sbb-select-forms',
+  templateUrl: './select-forms.component.html',
+  styleUrls: ['./select-forms.component.scss']
+})
+export class SelectFormsComponent {
+  value: string;
+  disabled = false;
+  optionDisabled = false;
+}