+++ projects/sbb-esta/angular-icons/schematics/svg-registry/schema.ts
@@ -1,0 +1,6 @@
+export interface Schema {
+  svgDir: string;
+  stripNumberIdInCollection: boolean;
+  stripNumberIds: boolean;
+  stripSbbInName: boolean;
+}